experimenting with writing trace addrs directly, fixing a bug with incorrect high interval

+++ lib/hplgst/hplgst.cc
@@ -23,7 +23,7 @@
 namespace __hplgst {
 
 void processRangeAccess(uptr PC, uptr Addr, uptr Size, bool IsWrite) {
-  /*VPrintf(3, "in hplgst::%s %p: %c %p %d\n", __FUNCTION__, PC,
+/*  VPrintf(3, "in hplgst::%s %p: %c %p %d\n", __FUNCTION__, PC,
           IsWrite ? 'w' : 'r', Addr, Size);*/
 
   // TODO do we need Size or PC?
@@ -47,7 +47,7 @@ void processRangeAccess(uptr PC, uptr Addr, uptr Size, bool IsWrite) {
     if (Addr - begin < m.interval_low())
       m.set_interval_low(Addr-begin);
     if (Addr-begin+Size > m.interval_high())
-      m.set_interval_high(Addr-begin+Size);
+      m.set_interval_high((u32)Addr-begin+Size);
 
     // this is prob too expensive
     // TODO make optional
